Also move util.{cc,h}
