+++ onnxruntime/contrib_ops/rocm/bert/fast_gelu_tunable_op.h
@@ -39,7 +39,7 @@ Status FastGeluOp(const FastGeluParams<T>* params) {
         (params->bias_length == 0 && params->input_length % VecSize == 0)));
 
   hipLaunchKernelGGL((FastGeluKernelVec<T, ThreadsPerBlock, VecSize>),
-                     dim3(CeilingDivision(params->input_length, ThreadsPerBlock * VecSize)),
+                     dim3(onnxruntime::rocm::tunable::CeilingDivision(params->input_length, ThreadsPerBlock * VecSize)),
                      dim3(ThreadsPerBlock),
                      0, params->stream,
                      params->input_length, params->bias_length, params->input, params->bias, params->output);
