+++ onnxruntime/core/providers/rocm/tunable/util.cc
@@ -1,11 +1,16 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "core/providers/rocm/tunable/tunable.h"
+#include "core/providers/rocm/tunable/util.h"
 
 namespace onnxruntime {
 namespace rocm {
 namespace tunable {
+
+int CeilingDivision(int n, int m) {
+  int r = (n - 1) / m + 1;
+  return r;
+}
 
 Timer::Timer(hipStream_t stream) : stream_(stream) {
   HIP_CHECK(hipEventCreate(&start_));
