+++ onnxruntime/core/providers/rocm/tunable/util.h
@@ -17,16 +17,30 @@
   } while (0)
 
 namespace onnxruntime {
-namespace contrib {
 namespace rocm {
+namespace tunable {
 
 int CeilingDivision(int n, int m);
 
-template<typename T, int VecSize>
+template <typename T, int VecSize>
 struct alignas(sizeof(T) * VecSize) AlignedVector {
   T val[VecSize];
 };
 
+class Timer {
+ public:
+  explicit Timer(hipStream_t stream);
+  void Start();
+  void End();
+  float Duration();
+  ~Timer();
+
+ private:
+  hipStream_t stream_;
+  hipEvent_t start_;
+  hipEvent_t end_;
+};
+
+}  // namespace tunable
 }  // namespace rocm
-}  // namespace contrib
 }  // namespace onnxruntime
