+++ onnxruntime/python/tools/kernel_explorer/kernel_explorer_interface.h
@@ -5,7 +5,7 @@
 
 #include <hip/hip_runtime.h>
 #include "core/providers/rocm/tunable/tunable.h"
-#include "contrib_ops/rocm/bert/util.h"
+#include "core/providers/rocm/tunable/util.h"
 
 using onnxruntime::rocm::tunable::Timer;
